Move sidebar fragments to own directory
